pyproject.toml
[project]
name = "ycm76229@gmail.com"
version = "1.1.0"
description = "django project"
authors = [
{name = "", email = ""},
]
license-expression = "MIT"
dependencies = [
    "django==3.2.4",
    "django-filter==2.4.0",
    "django-import-export>=2.8.0",
    "django-simpleui==2021.6.2",
    "djangorestframework==3.12.4",
    "djangorestframework-simplejwt==4.7.1",
    "drf-yasg==1.20.0",
    "psycopg2-binary",
    "pyyaml==5.4.1",
    "requests==2.25.1",
    "django-environ==0.4.5",
    "setuptools>=62.0.0",
    "gunicorn>=20.1.0",
]
requires-python = ">=3.8"

[project.urls]
Homepage = ""

[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "http://mirrors.aliyun.com/pypi/simple/"
verify_ssl = false

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
_.env_file = ".env"
mm = "python manage.py makemigrations"
mr = "python manage.py migrate"
mrs = { cmd = ["python","manage.py","migrate","--run-syncdb"]}
mmrs = { shell = "python manage.py makemigrations && python manage.py migrate --run-syncdb" }
gen = { shell = "python manage.py init_admin && python manage.py gen_all && cd tyadmin && npm install && npm run build" }
st = { cmd = ["python","manage.py","collectstatic", "-c"]}
cs = "python manage.py createsuperuser"
dev = "python manage.py runserver"
start = "gunicorn -w 8 -b localhost:8000 dj_server.wsgi"
